# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Books API",
      "description": "This is a books API",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:8085/api"
      }
    ],
    "tags": [
      {
        "name": "geolocation",
        "description": "All data about location you got"
      }
    ],
    "paths": {
      "/products": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Returns all data product",
          "description": "Returns all data product",
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Products"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "products"
          ],
          "summary": "Add a new product",
          "description": "Add a new product",
          "requestBody": {
            "description": "Create a new product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Products"
                  }
                }
              }
            },
            "405": {
              "description": "Invalid input"
            }
          }
        },
        "put": {
          "tags": [
            "products"
          ],
          "summary": "Change data product by id",
          "description": "Change data product by id",
          "requestBody": {
            "description": "Create a new product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Products"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Returns data trap by cid",
          "description": "Returns data trap by cid",
          "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "ID of product to return",
                "required": true,
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Products"
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "products"
          ],
          "summary": "Returns data trap by cid",
          "description": "Returns data trap by cid",
          "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "ID of product to return",
                "required": true,
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Products"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Products": {
          "required": [
            "id",
            "description",
            "nama",
            "price",
            "category_id"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": "1"
            },
            "product_descriptions": {
              "type": "string",
              "example": "We can describe some product in this field"
            },
            "product_name": {
              "type": "string",
              "example": "This field must be product name"
            },
            "price": {
              "type": "number",
              "example": "120000"
            },
            "category_id": {
              "type": "string",
              "example": "This filed is fill out of id of category"
            }
          }
        }
      },
      "requestBodies": {
        "TrapData": {
          "description": "Coordinate object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Products"
              }
            }
          }
        }
      }
    }
  }